Split ruler chart.
